third_party/WebKit/LayoutTests/external/wpt/css/css-transforms/parsing/perspective-origin-parsing-invalid-expected.txt

This is a testharness.js-based test.
PASS e.style['perspective-origin'] = "auto" should not set the property value
PASS e.style['perspective-origin'] = "1px 2px 3px" should not set the property value
PASS e.style['perspective-origin'] = "left right" should not set the property value
PASS e.style['perspective-origin'] = "bottom 10% top 20%" should not set the property value
FAIL e.style['perspective-origin'] = "center left 1px" should not set the property value assert_equals: expected "" but got "left 1px center"
FAIL e.style['perspective-origin'] = "center top 2px" should not set the property value assert_equals: expected "" but got "center top 2px"
FAIL e.style['perspective-origin'] = "right 3% center" should not set the property value assert_equals: expected "" but got "right 3% center"
FAIL e.style['perspective-origin'] = "left 4px top" should not set the property value assert_equals: expected "" but got "left 4px top"
FAIL e.style['perspective-origin'] = "right top 5px" should not set the property value assert_equals: expected "" but got "right top 5px"
FAIL e.style['perspective-origin'] = "bottom 6% center" should not set the property value assert_equals: expected "" but got "center bottom 6%"
FAIL e.style['perspective-origin'] = "bottom 7% left" should not set the property value assert_equals: expected "" but got "left bottom 7%"
FAIL e.style['perspective-origin'] = "bottom right 8%" should not set the property value assert_equals: expected "" but got "right 8% bottom"
Harness: the test ran to completion.
